requirements.txt
tensorflow
keras
pillow
numpy
opencv-python
flask
matplotlib